feat: more intelligent event dates in excerpt
Uses the Event Dates block template to render dates as part of Event listing excerpts. This is a better user experience because the dates don't have to be near the top of the listing content, and all dates will be pre-pended to the excerpt regardless of where they are in the event post content.